requirements.txt
#DRF
celery
django
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support
proj
GEOS
psycopg2
django-cors-headers